refactor(진성진): 아이템 페이지 로직 커스텀 훅 분리 (useItemPageState)
